requirements.txt
aiogram==3.22.0
dotenv==0.9.9
google-auth==2.40.3
gspread==6.2.1